+++ foundrytools_cli_2/__main__.py
@@ -2,6 +2,7 @@ import click
 
 from foundrytools_cli_2.cli.cff import cli as cff
 from foundrytools_cli_2.cli.converter import cli as converter
+from foundrytools_cli_2.cli.experimental import cli as experimental
 from foundrytools_cli_2.cli.fix import cli as fix
 from foundrytools_cli_2.cli.glyphs import cli as glyphs
 from foundrytools_cli_2.cli.gsub import cli as gsub
@@ -19,6 +20,7 @@ from foundrytools_cli_2.cli.utils import cli as utils
     commands={
         "cff": cff,
         "converter": converter,
+        "experimental": experimental,
         "fix": fix,
         "glyphs": glyphs,
         "gsub": gsub,
