+++ foundrytools_cli_2/cli/experimental/__init__.py
@@ -1,0 +1,3 @@
+from foundrytools_cli_2.cli.experimental.cli import cli
+
+__all__ = ["cli"]
