+++ foundrytools_cli_2/cli/experimental/cli.py
@@ -1,0 +1,36 @@
+# pylint: disable=import-outside-toplevel
+import typing as t
+from pathlib import Path
+
+import click
+
+from foundrytools_cli_2.cli.shared_options import base_options
+from foundrytools_cli_2.cli.task_runner import TaskRunner
+
+cli = click.Group(help="Experimental utilities.")
+
+
+@cli.command("recalc-stems")
+@base_options()
+def recalc_stems(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
+    """
+    Recalculates hinting stems for the given font files.
+    """
+    from foundrytools_cli_2.cli.experimental.tasks.recalc_stems import main as task
+
+    runner = TaskRunner(input_path=input_path, task=task, **options)
+    runner.filter.filter_out_tt = True
+    runner.run()
+
+
+@cli.command("recalc-zones")
+@base_options()
+def recalc_zones(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
+    """
+    Recalculates hinting zones for the given font files.
+    """
+    from foundrytools_cli_2.cli.experimental.tasks.recalc_zones import main as task
+
+    runner = TaskRunner(input_path=input_path, task=task, **options)
+    runner.filter.filter_out_tt = True
+    runner.run()
