+++ foundrytools_cli_2/cli/otf/cli.py
@@ -44,32 +44,6 @@ def dehint(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
     runner.run()
 
 
-@cli.command("recalc-stems")
-@base_options()
-def recalc_stems(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
-    """
-    Recalculates hinting stems for the given font files.
-    """
-    from foundrytools_cli_2.cli.otf.snippets.recalc_stems import main as task
-
-    runner = TaskRunner(input_path=input_path, task=task, **options)
-    runner.filter.filter_out_tt = True
-    runner.run()
-
-
-@cli.command("recalc-zones")
-@base_options()
-def recalc_zones(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
-    """
-    Recalculates hinting zones for the given font files.
-    """
-    from foundrytools_cli_2.cli.otf.snippets.recalc_zones import main as task
-
-    runner = TaskRunner(input_path=input_path, task=task, **options)
-    runner.filter.filter_out_tt = True
-    runner.run()
-
-
 @cli.command("subr")
 @base_options()
 def subr(input_path: Path, **options: t.Dict[str, t.Any]) -> None:
